solution: Remove Git submodules from project
This is for research and development and not a production project.

My thinking has changed. The submodule approach is very handy in a
group project context. This project is just for me. So I am going
to keep it simple. I have no reason to travel back through time with
this stuff. I have not really done so in nearly two years.

... so here we go!